Only attempt to delete the XPC targets folder if there is one
Thanks @jkmassel.

See https://github.com/Automattic/simplenote-macos/pull/824#discussion_r566577414

+++ Scripts/Build-Phases/sparkle-code-sign.sh
@@ -34,9 +34,10 @@ fi
 
 # Reset: Sparkle's Codesign fails when attempting to re-sign
 #
-echo "Nuking old XPC..."
-
-rm -r "${XPC_TARGET}"
+if [[ -d "${XPC_TARGET}" ]]; then
+  echo "Nuking old XPC..."
+  rm -r "${XPC_TARGET}"
+fi
 
 # Copy Phase
 #
